just small changes, still trying to show the availability in its own section
