+++ client/src/components/TDashboard/TDashboard.scss
@@ -1,0 +1,183 @@
+// .page-title {
+//   text-align: center;
+//   font-size: 2rem;
+//   margin-bottom: 1.5rem;
+// }
+
+// .container {
+//   display: flex;
+//   justify-content: space-between;
+//   flex-wrap: wrap;
+// }
+
+// .side-menu {
+//   width: 20%;
+//   padding: 1rem;
+//   border-right: 1px solid #ddd;
+// }
+
+// .profile-card {
+//   display: flex;
+//   align-items: center;
+//   padding: 1rem;
+//   margin-bottom: 2rem;
+//   border-bottom: 1px solid #ddd;
+//   border-radius: 5px;
+// }
+
+// .profile-img {
+//   width: 80px;
+//   height: 80px;
+//   border-radius: 50%;
+//   margin-right: 1rem;
+// }
+
+// .profile-name,
+// .profile-subject {
+//   display: block;
+// }
+
+// .profile-name {
+//   font-weight: bold;
+//   margin-bottom: 0.5rem;
+// }
+
+// .profile-subject {
+//   font-size: 0.9rem;
+//   color: #888;
+// }
+
+// .nav-link {
+//   display: flex;
+//   align-items: center;
+//   text-decoration: none;
+//   color: #333;
+//   margin-bottom: 0.5rem;
+// }
+
+// .nav-link i {
+//   margin-right: 0.5rem;
+// }
+
+// .nav-link:hover {
+//   color: #007bff;
+// }
+
+// .upcoming-appointments {
+//   width: 100%;
+//   padding: 1rem;
+// }
+
+// .calendar-container {
+//   margin-right: 2rem;
+// }
+
+// .calendar {
+//   width: 100%;
+//   padding: 1rem;
+// }
+
+// .submit-button {
+//   background-color: blue;
+//   color: white;
+//   border-radius: 20px;
+//   padding: 10px 20px;
+//   font-size: 16px;
+//   cursor: pointer;
+//   margin-top: 20px;
+//   border: none;
+//   text-align: center;
+//   display: block;
+//   width: 50%;
+//   margin-left: auto;
+//   margin-right: auto;
+// }
+
+// @media (min-width: 768px) {
+//   .side-menu {
+//     width: 20%;
+//     border-right: 1px solid #ddd;
+//     border-bottom: none;
+//   }
+//   .upcoming-appointments {
+//     width: 35%;
+//   }
+//   .calendar {
+//     width: 40%;
+//   }
+// }
+
+.pickdate {
+  min-height: 100vh;
+  .teachers-dashboard {
+    background-color: rgba(250, 184, 0, 0.924);
+    text-align: center;
+  }
+
+  .date-time {
+    padding: 2rem;
+
+    .profile-teacher {
+      border-right: 2px solid black;
+    }
+
+    .col-sm-8 {
+      padding: 2rem;
+      border-right: 2px solid black;
+      display: flex;
+      justify-content: center;
+    }
+    .react-calendar {
+      width: 500px;
+      margin: 1rem 0;
+      border-radius: 10px;
+      padding: 1rem;
+
+      .react-calendar__viewContainer {
+        text-align: center;
+        text-decoration-line: none;
+        font-weight: bolder;
+      }
+      .react-calendar__tile--now {
+        background: rgba(250, 184, 0, 0.4);
+        border-radius: 5px;
+        color: black;
+      }
+      .react-calendar__month-view__days__day:hover {
+        background-color: rgba(250, 184, 0, 0.924);
+        border-radius: 5px;
+      }
+      .react-calendar__tile--active {
+        background: rgba(250, 184, 0, 0.924);
+        color: black;
+        border-radius: 5px;
+      }
+      .react-calendar__tile--focus {
+        background: rgba(250, 184, 0, 0.924);
+        color: black;
+        border-radius: 5px;
+      }
+      .react-calendar__month-view__days__day--weekend {
+        color: #757575;
+      }
+
+      .time {
+        width: 150px;
+        margin: 1rem 0;
+      }
+
+      span {
+        font-weight: bolder;
+      }
+    }
+  }
+  .selected {
+    margin-top: 3rem;
+    text-align: center;
+  }
+
+  .upcoming-dates {
+    margin-top: 6rem;
+    text-align: center;
+  }
+}
